land: squash 4 attempt commits from afk/2132-extract-the-collection-contract-enforcer
- feat(contract): extract collection contract enforcer Refs #2132
- refactor(entity): delegate collection contract enforcement Refs #2132
- refactor(ports): re-export collection contract plans Refs #2132
- chore(application): register collection contract enforcer Refs #2132
